refactor(repos): 将 Cols 作为 base class，并使用类继承
