+++ os-checks/shared/repos.ts
@@ -1,3 +1,5 @@
+import { type Col, Cols } from "./columns-select"
+
 export type Info = {
   /// repo name
   name: string,
@@ -113,9 +115,13 @@ export type Repo = {
   topics: string[],
 }
 
-export type Col = { display: boolean, name: string, option: string };
-/// reactive Columns
-export const columns: { [key: string]: Col } = {
+
+const defaultColumns = [
+  "license", "homepage", "description", "pushed_at",
+  "active_days", "contributions", "contributors", "size",
+];
+
+const columns: { [key: string]: Col } = {
   license: { display: false, name: "License", option: "License" },
   homepage: { display: false, name: "Home", option: "Home Page" },
   description: { display: false, name: "Description", option: "Description" },
@@ -141,69 +147,9 @@ export const columns: { [key: string]: Col } = {
   topics: { display: false, name: "Topics", option: "Topics" },
 };
 
-export class Cols {
-  cols;
 
-  static full = "[Columns] Full";
-  static slimmed = "[Columns] Default (Slimmed)";
-
-  static defaultColumns = [
-    "license", "homepage", "description", "pushed_at",
-    "active_days", "contributions", "contributors", "size",
-  ];
-
-  constructor() {
-    this.cols = columns;
-  }
-
-  display(col: string) {
-    return this.cols[col].display;
-  }
-
-  name(col: string) {
-    return this.cols[col].name;
-  }
-
-  static options() {
-    return [Cols.full, Cols.slimmed, ...Object.keys(columns)];
-  }
-
-  static option(opt: string): string {
-    if (opt === Cols.full) {
-      return Cols.full;
-    } else if (opt === Cols.slimmed) {
-      return Cols.slimmed;
-    }
-    return columns[opt].option;
-  }
-
-  setDefaultColumns() {
-    const default_columns = new Set(Cols.defaultColumns);
-    for (const col of Object.keys(this.cols)) {
-      this.cols[col].display = default_columns.has(col);
-    }
-  }
-
-  setDisplay(cols: string[]) {
-    if (cols.length === 0) {
-      this.setDefaultColumns();
-      return;
-    }
-
-    if (cols.findIndex(c => c === Cols.full) !== -1) {
-      // display all columns 
-      for (const col of Object.keys(this.cols)) {
-        this.cols[col].display = true;
-      }
-      return;
-    }
-
-    const default_columns = (cols.findIndex(c => c === Cols.slimmed) !== -1) ?
-      Cols.defaultColumns : [];
-    const set = new Set([...cols, ...default_columns]);
-
-    for (const col of Object.keys(this.cols)) {
-      this.cols[col].display = set.has(col);
-    }
+export class RepoCols extends Cols {
+  static init() {
+    return new RepoCols(columns, defaultColumns);
   }
 }
